Fix issues with FUNCALL symbol.
When CL:FUNCALL is applied on a symbol rather than a function object,
the function in the symbol's SYMBOL-FUNCTION is called. That is the
global function definition is called and not a lexical function
definition. This means (FUNCALL 'FOO ...) is not always equivalent
to (FOO ...), if there is a lexical definition for FOO, defined by
FLET or LABELS, in the environment.

This commit adds checks for whether there is a lexical definition for
a function of the same name, and only if there isn't is the
optimization performed.